requirements.txt
# Core dependencies
streamlit>=1.40.2
python-dotenv>=1.0.1

# AI/ML Models - Gemini & Gemma
google-generativeai>=0.8.3
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99
sentence-transformers>=2.7.0

# Vector Search
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
numpy>=1.26.4

# Data Processing
pandas>=2.2.3
jsonlines>=4.0.0
tqdm>=4.66.0
scikit-learn>=1.5.0

# PDF Generation
reportlab>=4.2.5
pylatex>=1.4.2
pypdf>=3.17.0
matplotlib>=3.9.2
sympy>=1.12

# Web and API
requests>=2.31.0
aiohttp>=3.9.0

# Development tools
pytest>=7.4.0
black>=23.0.0
pylint>=3.0.0